Moved extensions from the Microsoft.Toolkit.Uwp.UI namespace to  Microsoft.Toolkit.Uwp.UI.Extensions
