+++ Microsoft.Toolkit.Uwp.UI/Extensions/HyperlinkExtensions.cs
@@ -14,7 +14,7 @@ using System.Windows.Input;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Documents;
 
-namespace Microsoft.Toolkit.Uwp.UI
+namespace Microsoft.Toolkit.Uwp.UI.Extensions
 {
     /// <summary>
     /// Provides attached dependency properties for the <see cref="Hyperlink"/> content element that allows
